refactoring components refactoring component. layout fixed.
