--- src/components/BullButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './BullButton'
