+++ src/components/CricketNumberCountSettingModal/CricketNumberCountSettingModal.test.tsx
@@ -1,0 +1,7 @@
+import { render } from '@testing-library/react'
+import CricketNumberCountSettingModal from './CricketNumberCountSettingModal'
+
+test('should rendering', () => {
+  const { container } = render(<CricketNumberCountSettingModal />)
+  expect(container).toMatchSnapshot()
+})
