+++ src/components/CricketNumberCountSettingModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from './CricketNumberCountSettingModal'
