+++ src/components/Footer/Footer.test.tsx
@@ -3,6 +3,6 @@ import Footer from './Footer'
 
 test('should rendering', () => {
   const { container } = render(<Footer />)
-  expect(screen.getAllByText('(c)2022-2022 subaru-git')).toHaveLength(1)
+  expect(screen.getAllByText('(c) 2022 okadat')).toHaveLength(1)
   expect(container).toMatchSnapshot()
 })
