+++ src/components/Footer/Footer.tsx
@@ -1,22 +1,25 @@
 import React, { FC } from 'react'
-import { css } from '@emotion/react'
+import { Box, Center, HStack, Link, Text } from '@chakra-ui/react'
+import { AiFillGithub, AiOutlineTwitter } from 'react-icons/ai'
 
 const Footer: FC = () => {
   return (
-    <div css={styles}>
-      <p>{'(c)2022-2022 subaru-git'}</p>
-    </div>
+    <Box bg='gray.700' h='24px' w='100%' bottom={0} position='absolute' mb='-32px'>
+      <Center>
+        <HStack gap={8}>
+          <Text color='white'>{`(c) 2022 okadat`}</Text>
+          <HStack gap={2}>
+            <Link href='https://github.com/subaru-git'>
+              <AiFillGithub size='24px' color='white' />
+            </Link>
+            <Link href='https://twitter.com/subaru_m_'>
+              <AiOutlineTwitter size='24px' color='white' />
+            </Link>
+          </HStack>
+        </HStack>
+      </Center>
+    </Box>
   )
 }
-const styles = css({
-  position: 'absolute',
-  bottom: 0,
-  left: 0,
-  backgroundColor: 'black',
-  textAlign: 'center',
-  color: 'white',
-  height: '32px',
-  width: '100%',
-  marginBottom: '-32px',
-})
+
 export default Footer
