+++ src/components/TargetBoard/TargetBoard.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react'
-import { Heading } from '@chakra-ui/react'
+import { Box, Heading } from '@chakra-ui/react'
 
 type TargetBoardProps = {
   target?: string
@@ -9,14 +9,14 @@ type TargetBoardProps = {
 
 const TargetBoard: FC<TargetBoardProps> = ({ target, message, size = 'md' }) => {
   return (
-    <>
+    <Box width={300}>
       <Heading fontSize={size === 'md' ? '32px' : '24px'} color='green.500' pl={4} pt={4}>
         {message}
       </Heading>
       <Heading textAlign='center' fontSize={size === 'md' ? '96px' : '64px'} py={8}>
         {target}
       </Heading>
-    </>
+    </Box>
   )
 }
